# openapi.yaml
openapi: 3.1.0
info:
  title: Memoria API
  version: 1.0.0
  description: Private photo/video sharing API (Capsules + Albums). All times UTC in JSON.
servers:
  - url: http://localhost:8080
    description: Local docker-compose
  - url: https://api.memoria.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
paths:
...
  /v1/me/notification-prefs:
    get:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: prefs array }
    put:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Updated prefs }
...
  /v1/media/presign:
    post:
      tags: [media]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, content_type, byte_size]
              properties:
                kind: { type: string, enum: [photo, video, voice] }
                content_type: { type: string }
                byte_size: { type: integer }
                duration_ms: { type: integer }
      responses:
        '200': { description: media_id + upload_url (10 min TTL) }
  /v1/media/{id}/confirm:
    post:
      tags: [media]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Ready media DTO }
  /v1/media/{id}:
    get:
      tags: [media]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Media with signed url (10 min TTL) }
...
  /v1/notifications/unread-count:
    get:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: count }
  /v1/notifications:
    get:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, maximum: 50 }
      responses:
        '200': { description: Notification feed }
  /v1/notifications/read-all:
    post:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: All marked read }
  /v1/notifications/{id}/read:
    post:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200': { description: Marked read }
  /v1/notifications/flush:
    post:
      tags: [notifications]
      security: [{ bearerAuth: [] }]
      summary: Flush due batches on app open
      responses:
        '200': { description: flushed }
...